Added upload parts from CSV and gzip files
Some of these functions can probably be combined. One reads a normal CSV file and streams it to Domo. Another reads a normal CSV, compresses it to gzip, and then uploads it to Domo (I am not sure if this would increase speed or reduce network requirements). And finally, one function reads a gzip file and uploads gzip parts to Domo.