feat(auth): Add scope support (#101)
Adds support for specifying `scope` when creating `Domo` SDK client.
`scope` is a list of scope names such as `["data", "user", "account"]`.

This allows limiting the scope of access tokens to a subset of the
scopes granted on the API client. Not specifying `scope` maintains the
default behavior of access tokens having the same scopes as the API
client.