bugfix in UtilitiesClient.stream_upload
In UtilitiesClient.stream_upload, the domo stream upload part api expects part_id to be contiguous. It seems to tolerate some discontinuities, but when uploading a large dataset the api eventually returns an error that there are too many gaps in the part_ids. This commit fixes this by using `i` as the part_id (instead of `start`) which will be contiguous.