Adding connection_file as a parameter for Domo init method
This is an ini file that contains all parameters the SDK uses to connect
to Domo (client_id, client_secret, and optionally api_host and
use_https). An example connection.ini file is provided in the examples
directory. Users may still a provide client_id and client_secret, but
they cannot provide both a client_id/secret and a connection_file.